Added Authentication page and component
